Corrigido teste de handler e nomenclatura de estruturas.

+++ domain/account.go
@@ -9,6 +9,6 @@ type Account struct {
 	Id        bson.ObjectId `json:"id" bson:"_id,omitempty"`
 	Name      string        `json:"name" bson:"name"`
 	Cpf       string        `json:"cpf" bson:"cpf"`
-	Ballance  int           `json:"ballance" bson:"ballance"`
+	Ballance  float64       `json:"ballance" bson:"ballance"`
 	CreatedAt time.Time     `json:"created_at" bson:"created_at"`
 }
